I might have to break this file out more but, so far, this is much nicer than before.
Beyond just stacking everything into the same function, I think not
relying on rebuilding on object to mean recalling the object (which
_also_ happens to store ALL the other functions for said object
nestled in there which makes for a headache to develop on or update
said object) helps, along with not requiring the columns to constantly
store representations of each row of themselves.

That all makes it much easier to just segment different actions into
different functions.

The old way had everything touching everything else and, already,
storing the drawing of the window in its own function that only cares
about the info./data stored in the `column-window` object makes it
much easier to focus on each particular task for each function and not
cross tasks all the time.

I can create a `rebuild` function and trust, so long as I get the
necessary data correct, the render function will redraw the window
just fine on its own and I don't have to keep tweaking the logic in
both places consistently.

Redrawing the entire window for the render function, regardless
whatever performance hit (if even any) is incurred, has resulted in
far less weird render issues this time around, too.