Object to handle storing Play Window data.
Because updates for the window are being run every second in a
separate thread, I don't want to risk introducing state so we're
avoiding the records for the object, this time.

Some utility functions to handle processing the list properly masks
most of any difficulty that might otherwise be encountered in the illusion.