And the logic for the Play Window; I think I simplified a lot of the original logic (though there were so many places to start…).
Now the logic is focused into a function to write the first line, a
function to write the second, a function for rebuilding the
window (which is mostly a wrapper of the render function), and a
function to render the window.

The render function does pack a lot of various logic into it but it's
also able to be used as the primary update function so I don't have to
split the logic over a million different places.