This is, honestly, a shim (and not even one that fully solves the problem) but it should guarantee that, mostly, the app. doesn't crash from calling MPD so often during intensive rendering (which is really just if a user were to rapidly resize their terminal window for an extended period of time).
I want to find a real solution to this but that may be just solving
the dropped connection issue of the MPD library or writing it in a way
that doesn't force that solve onto the user.

Or maybe updating `main-loop` to handle processing keys better (such
that it doesn't keep scrolling pages down after you release `pg up` or
`pg dn`); I dunno. I'll solve it later.