Make use of the new `seek` function.
The setup I made for catching modifier keys aren't working for
urxvt (which I guess I should have realized it wasn't necessarily work
for every possible terminal) plus I learned about `KEY_SRIGHT` and
`KEY_SLEFT` so let's make use of those and change the bindings a bit.

I'll have to figure out what to do with what was originally going to
be using the `ctrl` key but it's not the end of the world to use
different bindings that, at least, will be guaranteed to work.